added register service to RegisterForm
