+++ src/services/register.js
@@ -2,7 +2,7 @@ import axios from 'axios'
 
 import { API_URL } from '../config'
 
-export const signup = async (user) => {
+export const register = async (user) => {
   try {
     const { data } = await axios.post(`${API_URL}/users/signup`, user)
